+++ Example/SwiftEntryKit/Presets/PresetsViewController.swift
@@ -284,9 +284,15 @@ class PresetsViewController: UIViewController {
         SwiftEntryKit.display(entry: contentView, using: attributes)
     }
     
-    // Bumps a custom nib view
+    // Bumps a custom nib originated view
     private func showCustomNibView(attributes: EKAttributes) {
         SwiftEntryKit.display(entry: NibExampleView(), using: attributes)
+    }
+    
+    // Bumps a custom view controller that is using a view from nib
+    private func showCustomViewController(attributes: EKAttributes) {
+        let viewController = ExampleViewController(with: NibExampleView())
+        SwiftEntryKit.display(entry: viewController, using: attributes)
     }
     
     // Sign in form
@@ -433,6 +439,8 @@ extension PresetsViewController {
             showButtonBarMessage(attributes: attributes)
         case 4:
             showAlertView(attributes: attributes)
+        case 5:
+            showRatingView(attributes: attributes)
         default:
             break
         }
@@ -456,7 +464,7 @@ extension PresetsViewController {
         case 0:
             showCustomNibView(attributes: attributes)
         case 1:
-            showRatingView(attributes: attributes)
+            showCustomViewController(attributes: attributes)
         default:
             break
         }